+++ contract/src/template.compact
@@ -433,7 +433,7 @@ pure circuit add_mod(start: Uint<32>, add: Uint<32>, n: Uint<32>): Uint<32> {
     return (start + add - (((start + add >= n) as Uint<1>) * n)) as Uint<32>;
 }
 
-export circuit combat_round(battle_id: Field): Maybe<BattleRewards> {
+export circuit combat_round(battle_id: Field, ability_targets: Maybe<Vector<3, Uint<0..2>>>): Maybe<BattleRewards> {
     const player_pub_key = derive_player_pub_key(disclose(player_secret_key()));
     assert(player_pub_key == active_battle_configs.lookup(disclose(battle_id)).player_pub_key, "Player auth failed");
 
@@ -446,7 +446,8 @@ export circuit combat_round(battle_id: Field): Maybe<BattleRewards> {
         all_abilities.lookup(index<Field>(loadout.abilities, state.deck_indices[2])),
     ];
     
-    const result = combat_round_impl(disclose(battle_id), round_abilities, mock_rand_target());
+    const result = combat_round_impl(disclose(battle_id), round_abilities, disclose(ability_targets));
+    
     active_battle_states.insert(disclose(battle_id), result);
     if (result.player_hp == 0) {
         active_battle_configs.remove(disclose(battle_id));
@@ -481,53 +482,6 @@ export circuit combat_round(battle_id: Field): Maybe<BattleRewards> {
     return none<BattleRewards>();
 }
 
-export circuit combat_round_with_targets(battle_id: Field, ability_targets: Vector<3, Uint<0..2>>): Maybe<BattleRewards> {
-    const player_pub_key = derive_player_pub_key(disclose(player_secret_key()));
-    assert(player_pub_key == active_battle_configs.lookup(disclose(battle_id)).player_pub_key, "Player auth failed");
-
-    const config = disclose(active_battle_configs.lookup(disclose(battle_id)));
-    const loadout = config.loadout;
-    const state = disclose(active_battle_states.lookup(disclose(battle_id)));
-    const round_abilities = [
-        all_abilities.lookup(index<Field>(loadout.abilities, state.deck_indices[0])),
-        all_abilities.lookup(index<Field>(loadout.abilities, state.deck_indices[1])),
-        all_abilities.lookup(index<Field>(loadout.abilities, state.deck_indices[2])),
-    ];
-    
-    const result = combat_round_impl_with_targets(disclose(battle_id), round_abilities, disclose(ability_targets));
-    active_battle_states.insert(disclose(battle_id), result);
-    if (result.player_hp == 0) {
-        active_battle_configs.remove(disclose(battle_id));
-        active_battle_states.remove(disclose(battle_id));
-        
-        return some<BattleRewards>(BattleRewards {
-            false,
-            0,
-            none<Field>()
-        });
-    } else if (result.enemy_hp_0 == 0 && result.enemy_hp_1 == 0 && result.enemy_hp_2 == 0) {
-        return_loadout(loadout);
-        
-        if (config.stats[0].boss_type != BOSS_TYPE.normal || (config.enemy_count >= 2 && config.stats[1].boss_type != BOSS_TYPE.normal) || (config.enemy_count >= 3 && config.stats[2].boss_type != BOSS_TYPE.normal)) {
-            const reward_ability = random_ability(2);
-            player_abilities.lookup(player_pub_key).insert(reward_ability, (player_abilities.lookup(player_pub_key).lookup(reward_ability) + 1) as Uint<32>);
-        }
-
-        const reward = BattleRewards {
-            true,
-            100,
-            none<Field>()
-        };
-        const player_stats = players.lookup(derive_player_pub_key(disclose(player_secret_key())));
-
-        players.insert(derive_player_pub_key(disclose(player_secret_key())), add_player_rewards(player_stats, reward));
-        active_battle_configs.remove(disclose(battle_id));
-        active_battle_states.remove(disclose(battle_id));
-
-        return some<BattleRewards>(reward);
-    }
-    return none<BattleRewards>();
-}
 
 pure circuit add_player_rewards(player: Player, reward: BattleRewards): Player {
     return Player {
@@ -535,40 +489,16 @@ pure circuit add_player_rewards(player: Player, reward: BattleRewards): Player {
     };
 }
 
-circuit combat_round_impl_with_targets(battle_id: Field, abilities: Vector<3, Ability>, ability_targets: Vector<3, Uint<0..2>>): BattleState {
+circuit combat_round_impl(battle_id: Field, abilities: Vector<3, Ability>, ability_targets: Maybe<Vector<3, Uint<0..2>>>): BattleState {
     const battle = active_battle_configs.lookup(battle_id);
     const old_state = active_battle_states.lookup(battle_id);
 
-    // Modified code generation needed to use ability_targets instead of random targeting
-    // For now, use the first target as fallback until full implementation
-    const target = ability_targets[0] as Uint<32>;
-
-    INSERT_ENEMY_BLOCK_CODE_HERE
-
-    INSERT_PLAYER_BLOCK_CODE_HERE
-
-    INSERT_PLAYER_DAMAGE_CODE_HERE
-
-    const new_enemy_dmg_0 = enemy_block_0 < player_damage_0 ? (old_state.enemy_hp_0 + player_damage_0 - enemy_block_0) as Uint<32> : 0;
-    const new_enemy_dmg_1 = enemy_block_1 < player_damage_1 ? (old_state.enemy_hp_1 + player_damage_1 - enemy_block_1) as Uint<32> : 0;
-    const new_enemy_dmg_2 = enemy_block_2 < player_damage_2 ? (old_state.enemy_hp_2 + player_damage_2 - enemy_block_2) as Uint<32> : 0;
-
-    INSERT_ENEMY_DAMAGE_CODE_HERE
-
-    INSERT_DECK_INDEX_CALCULATION_CODE_HERE
-
-    return BattleState {
-        INSERT_DECK_INDEX_BATTLE_STATE_INIT_CODE_HERE,
-        player_block < enemy_damage ? (old_state.player_hp + enemy_damage - player_block) as Uint<32> : 0,
-        new_enemy_dmg_0,
-        new_enemy_dmg_1,
-        new_enemy_dmg_2
+    // Handle targeting: use player targets if provided, otherwise random target
+    const target = if (ability_targets.is_some) {
+        ability_targets.value[0] as Uint<32>  // For now, use first target as fallback
+    } else {
+        mock_rand_target()
     };
-}
-
-circuit combat_round_impl(battle_id: Field, abilities: Vector<3, Ability>, target: Uint<32>): BattleState {
-    const battle = active_battle_configs.lookup(battle_id);
-    const old_state = active_battle_states.lookup(battle_id);
 
     // See generate.js for an explanation of the code generation
